+++ CHANGELOG.md
@@ -6,7 +6,8 @@ Development version
 
 New features and improvements:
 
-  - Add where filter to ``read_file`` (#2552)
+- Added ``normalize()`` method from shapely to GeoSeries/GeoDataframe (#2537)
+- Add where filter to ``read_file`` (#2552)
 
 Deprecations and compatibility notes: